fix: current and prev list comparisons of menu lists
